feat: Add ScriptProperty for GroupId in PSSymantecSEPM.Types.ps1xml
This commit adds a new ScriptProperty called GroupId to the PSSymantecSEPM.Types.ps1xml file. The GroupId property retrieves the ID of the group and returns it. This enhancement improves the functionality of the file by providing an easy way to access the group ID.

feat: Update Get-SEPMLocations to accept GroupID via pipeline

This commit modifies the Get-SEPMLocations function in the Get-SEPMLocations.ps1 file to accept the GroupID parameter via pipeline. This enhancement allows users to directly pipe the output of Get-SEPMGroups to Get-SEPMLocations, making it more convenient to retrieve location names and IDs for all groups.